+++ src/game/scenes/world/entities/addons/__tests__/live.test.ts
