+++ src/game/scenes/world/entities/addons/indicator.ts
