+++ src/game/scenes/world/entities/npc/variants/enemy/enemy.ts
@@ -11,8 +11,8 @@ import { LEVEL_TILE_SIZE } from '~const/world/level';
 import { Building } from '~entity/building';
 import { NPC } from '~entity/npc';
 import { Assets } from '~lib/assets';
-import { progressionLinear, progressionQuadratic } from '~lib/difficulty';
-import { excludePosition } from '~lib/utils';
+import { progressionLinear, progressionQuadratic } from '~lib/progression';
+import { excludePosition } from '~lib/dimension';
 import { Effect, Particles } from '~scene/world/effects';
 import { Level } from '~scene/world/level';
 import { GameFlag, GameSettings } from '~type/game';
