--- src/lib/__mocks__/positions.ts
@@ -1,10 +1,0 @@
-import { Vector2D } from '~type/world/level';
-
-const positions: Vector2D[] = [
-  { x: 0, y: 0 },
-  { x: 10, y: 10 },
-  { x: 20, y: -10 },
-  { x: 10, y: 0 },
-];
-
-export default positions;
