+++ src/lib/progression.ts
